+++ lib/src/realm_class.dart
@@ -96,6 +96,7 @@ export 'set.dart' show RealmSet, RealmSetChanges, RealmSetOfObject;
 export 'migration.dart' show Migration;
 export 'realm_object.dart'
     show
+        AsymmetricObject,
         DynamicRealmObject,
         EmbeddedObject,
         EmbeddedObjectExtension,
@@ -276,6 +277,19 @@ class Realm implements Finalizable {
     object.manage(this, handle, accessor, update);
 
     return object;
+  }
+
+  /// Ingest an [AsymmetricObject] to the [Realm].
+  ///
+  /// Ingesting is a write only operation. The ingested objects synchronizes to
+  /// the App Services backend and are deleted from the device. An [AsymmetricObject]
+  /// can never be read from the Realm.
+  void ingest<T extends AsymmetricObject>(T object) {
+    final metadata = _metadata.getByType(object.runtimeType);
+    final handle = _createObject(object, metadata, false);
+
+    final accessor = RealmCoreAccessor(metadata, _isInMigration);
+    object.manage(this, handle, accessor, false);
   }
 
   RealmObjectHandle _createObject(RealmObjectBase object, RealmObjectMetadata metadata, bool update) {