+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Enhancements
 * Support efficient `skip` on `RealmResults` ([#1391](https://github.com/realm/realm-dart/pull/1391))
 * Support efficient `indexOf` and `contains` on `RealmResults` ([#1394](https://github.com/realm/realm-dart/pull/1394))
+* Support asymmetric objects. ([#1400](https://github.com/realm/realm-dart/pull/1400))
+
 ### Fixed
 * None